examples/fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
pydantic==2.4.2
pyjwt==2.8.0
python-multipart==0.0.6
aicore @ git+https://github.com/BrunoV21/AiCore@0.1.9